RubyDialogCallback_drawEllipse is modified so as to avoid use of 'rad1' and 'rad2' arguments (these symbols are reserved for MFC in Windows)
git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/molby/trunk@408 a2be9bc6-48de-4e38-9406-05402d4bc13c

+++ wxSources/MyListCtrl.cpp
@@ -142,11 +142,11 @@ MyListCtrl::OnGetItemAttr(long item) const
 	if (ret == 0)
 		return NULL;
 	if (ret & 1) {
-		wxColour fgcol(fg[0] * 255, fg[1] * 255, fg[2] * 255);
+		wxColour fgcol((int)(fg[0] * 255), (int)(fg[1] * 255), (int)(fg[2] * 255));
 		subTitleRowAttr->SetTextColour(fgcol);
 	} else subTitleRowAttr->SetTextColour(*wxBLACK);
 	if (ret & 2) {
-		wxColour bgcol(bg[0] * 255, bg[1] * 255, bg[2] * 255);
+		wxColour bgcol((int)(bg[0] * 255), (int)(bg[1] * 255), (int)(bg[2] * 255));
 		subTitleRowAttr->SetBackgroundColour(bgcol);
 	} else subTitleRowAttr->SetBackgroundColour(*wxWHITE);
 	return subTitleRowAttr;
